[ZP-178] fix noteHadler, several rest methods

+++ zeppelin-server/src/main/java/org/apache/zeppelin/websocket/handler/NoteHandler.java
@@ -17,13 +17,12 @@
 
 package org.apache.zeppelin.websocket.handler;
 
-import java.io.File;
 import java.io.IOException;
+import java.nio.file.AccessDeniedException;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.apache.commons.lang3.StringUtils;
@@ -33,8 +32,6 @@ import org.apache.zeppelin.websocket.ConnectionManager;
 import org.apache.zeppelin.websocket.Operation;
 import org.apache.zeppelin.websocket.SockMessage;
 import org.apache.zeppelin.websocket.dto.NoteDTOConverter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.WebSocketSession;
@@ -49,7 +46,6 @@ import ru.tinkoff.zeppelin.engine.NoteService;
 @Component
 public class NoteHandler extends AbstractHandler {
 
-  private static final Logger LOG = LoggerFactory.getLogger(NoteHandler.class);
   private static final String TRASH_FOLDER = "~Trash";
 
   private final NoteDTOConverter noteDTOConverter;
@@ -63,23 +59,11 @@ public class NoteHandler extends AbstractHandler {
   }
 
 
-  public void listNotesInfo(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
-    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
-    final List<Note> notes = noteService.getAllNotes();
-
-    final Set<String> admin = new HashSet<>();
-    admin.addAll(Configuration.getAdminUsers());
-    admin.addAll(Configuration.getAdminGroups());
-
-    final Set<String> userRoles = new HashSet<>();
-    userRoles.addAll(authenticationInfo.getRoles());
-    userRoles.add(authenticationInfo.getUser());
-
-    final List<NoteInfo> notesInfo = notes
-            .stream()
-            .filter(note -> new HashSet<>(userRoles).removeAll(note.getReaders()) || new HashSet<>(admin).removeAll(userRoles))
-            .map(NoteInfo::new)
-            .collect(Collectors.toList());
+  public void sendListNotesInfo(final WebSocketSession conn) throws IOException {
+    final List<NoteInfo> notesInfo = noteService.getAllNotes().stream()
+        .filter(this::userHasOwnerPermission)
+        .map(NoteInfo::new)
+        .collect(Collectors.toList());
 
     conn.sendMessage(new SockMessage(Operation.NOTES_INFO).put("notes", notesInfo).toSend());
   }
@@ -111,30 +95,32 @@ public class NoteHandler extends AbstractHandler {
   public void updateNote(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
     final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
     final Note note = safeLoadNote("id", fromMessage, Permission.READER, authenticationInfo, conn);
-    final String path = fromMessage.getNotNull("path");
-    final Map config = fromMessage.getOrDefault("config", null);
+    final String path = normalizePath(fromMessage.getNotNull("path"));
 
     note.setPath(path);
-
     noteService.updateNote(note);
+    sendListNotesInfo(conn);
   }
 
   public void deleteNote(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
     final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
 
-    final Note note = safeLoadNote("id", fromMessage, Permission.OWNER, authenticationInfo, conn);
+    final Note note = noteService.getNote((String) fromMessage.getNotNull("id"));
+    if (!userHasOwnerPermission(note)) {
+      throw new AccessDeniedException(
+          "User " + authenticationInfo.getUser() +
+              " is no owner note " + note.getId()
+      );
+    }
     noteService.deleteNote(note);
-
     connectionManager.removeNoteSubscribers(note.getId());
+    sendListNotesInfo(conn);
   }
 
   public void createNote(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
     final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
 
-    String notePath = fromMessage.getNotNull("path");
-    if (!notePath.startsWith(File.separator)) {
-      notePath = File.separator + notePath;
-    }
+    String notePath = normalizePath(fromMessage.getNotNull("path"));
 
     try {
       final Note note = new Note(notePath);
@@ -161,6 +147,7 @@ public class NoteHandler extends AbstractHandler {
 
       connectionManager.addSubscriberToNode(note.getId(), conn);
       conn.sendMessage(new SockMessage(Operation.NEW_NOTE).put("note", note).toSend());
+      sendListNotesInfo(conn);
     } catch (final Exception e) {
       throw new IllegalStateException("Failed to create note.", e);
     }
@@ -170,10 +157,7 @@ public class NoteHandler extends AbstractHandler {
     final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
 
     final Note note = safeLoadNote("id", fromMessage, Permission.READER, authenticationInfo, conn);
-    String path = fromMessage.getNotNull("name");
-    if (!path.startsWith(File.separator)) {
-      path = File.separator + path;
-    }
+    String path = normalizePath(fromMessage.getNotNull("name"));
 
     Note cloneNote = new Note(path);
     cloneNote.setPath(path);
@@ -212,17 +196,14 @@ public class NoteHandler extends AbstractHandler {
 
   public void moveNoteToTrash(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
     final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
-
     final Note note = safeLoadNote("id", fromMessage, Permission.OWNER, authenticationInfo, conn);
-
-    note.setPath("/" + TRASH_FOLDER + note.getPath());
-
+    note.setPath("/" + TRASH_FOLDER + normalizePath(note.getPath()));
     noteService.updateNote(note);
+    sendListNotesInfo(conn);
   }
 
   public void restoreNote(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
     final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
-
     final Note note = safeLoadNote("id", fromMessage, Permission.OWNER, authenticationInfo, conn);
 
     if (!note.getPath().startsWith("/" + TRASH_FOLDER)) {
@@ -230,162 +211,88 @@ public class NoteHandler extends AbstractHandler {
     }
 
     final String destNotePath = note.getPath().replace("/" + TRASH_FOLDER, "");
-
-    note.setPath(destNotePath);
+    note.setPath(normalizePath(destNotePath));
     noteService.updateNote(note);
+    sendListNotesInfo(conn);
   }
 
-  public void restoreFolder(final SockMessage fromMessage) throws IOException {
-    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
+  public void restoreFolder(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
+    final String folderPath = normalizePath(fromMessage.getNotNull("id")) + "/";
 
-    final String folderPath = "/" + fromMessage.getNotNull("id") + "/";
     if (!folderPath.startsWith("/" + TRASH_FOLDER)) {
       throw new IOException("Can't restore folder: '" + folderPath + "' as it is not in trash folder");
     }
-    final Set<String> admin = new HashSet<>();
-    admin.addAll(Configuration.getAdminUsers());
-    admin.addAll(Configuration.getAdminGroups());
 
-    final Set<String> userRoles = new HashSet<>();
-    userRoles.addAll(authenticationInfo.getRoles());
-    userRoles.add(authenticationInfo.getUser());
-
-    final List<Note> notes = noteService.getAllNotes()
-            .stream()
-            .filter(note -> !note.getPath().startsWith(folderPath))
-            .filter(note -> new HashSet<>(userRoles).removeAll(note.getReaders()) || new HashSet<>(admin).removeAll(userRoles))
-            .collect(Collectors.toList());
-    //TODO: add check for permission
-
-    for (final Note note : notes) {
-      final String notePath = note.getPath().substring(TRASH_FOLDER.length() + 1);
-      note.setPath(notePath);
-      noteService.updateNote(note);
-    }
+    noteService.getAllNotes().stream()
+        .filter(this::userHasOwnerPermission)
+        .filter(note -> note.getPath().startsWith(folderPath))
+        .forEach(note -> {
+          String notePath = normalizePath(note.getPath().substring(TRASH_FOLDER.length() + 1));
+          note.setPath(notePath);
+          noteService.updateNote(note);
+        });
+    sendListNotesInfo(conn);
   }
 
 
-  public void renameFolder(final SockMessage fromMessage) throws IOException {
-    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
+  public void renameFolder(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
+    final String oldFolderPath = normalizePath(fromMessage.getNotNull("id")) + "/";
+    final String newFolderPath = normalizePath(fromMessage.getNotNull("name")) + "/";
 
-    final String oldFolderPath = "/" + fromMessage.getNotNull("id");
-    final String newFolderPath = "/" + fromMessage.getNotNull("name");
-
-    final Set<String> admin = new HashSet<>();
-    admin.addAll(Configuration.getAdminUsers());
-    admin.addAll(Configuration.getAdminGroups());
-
-    final Set<String> userRoles = new HashSet<>();
-    userRoles.addAll(authenticationInfo.getRoles());
-    userRoles.add(authenticationInfo.getUser());
-
-    final List<Note> notes = noteService.getAllNotes()
-            .stream()
-            .filter(note -> !note.getPath().startsWith(oldFolderPath + "/"))
-            .filter(note -> new HashSet<>(userRoles).removeAll(note.getReaders()) || new HashSet<>(admin).removeAll(userRoles))
-            .collect(Collectors.toList());
-
-    for (final Note note : notes) {
-      final String notePath = note.getPath().replaceFirst(oldFolderPath, newFolderPath);
-      note.setPath(notePath);
-      noteService.updateNote(note);
-    }
+    noteService.getAllNotes().stream()
+        .filter(this::userHasOwnerPermission)
+        .filter(note -> note.getPath().startsWith(oldFolderPath))
+        .forEach(note -> {
+          String notePath =
+              normalizePath(note.getPath().replaceFirst(oldFolderPath, newFolderPath));
+          note.setPath(notePath);
+          noteService.updateNote(note);
+        });
+    sendListNotesInfo(conn);
   }
 
-  public void moveFolderToTrash(final SockMessage fromMessage) throws IOException {
-    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
-    final String folderPath = "/" + fromMessage.getNotNull("id") + "/";
+  public void moveFolderToTrash(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
+    final String folderPath = normalizePath(fromMessage.getNotNull("id")) + "/";
 
-    final Set<String> admin = new HashSet<>();
-    admin.addAll(Configuration.getAdminUsers());
-    admin.addAll(Configuration.getAdminGroups());
-
-    final Set<String> userRoles = new HashSet<>();
-    userRoles.addAll(authenticationInfo.getRoles());
-    userRoles.add(authenticationInfo.getUser());
-
-    final List<Note> notes = noteService.getAllNotes()
-            .stream()
-            .filter(note -> !note.getPath().startsWith(folderPath))
-            .filter(note -> new HashSet<>(userRoles).removeAll(note.getReaders()) || new HashSet<>(admin).removeAll(userRoles))
-            .collect(Collectors.toList());
-
-    for (final Note note : notes) {
-      final String notePath = "/" + TRASH_FOLDER + note.getPath();
-      note.setPath(notePath);
-      noteService.updateNote(note);
-    }
+    noteService.getAllNotes().stream()
+        .filter(note -> note.getPath().startsWith(folderPath))
+        .filter(this::userHasOwnerPermission)
+        .forEach(note -> {
+          String notePath = "/" + TRASH_FOLDER + normalizePath(note.getPath());
+          note.setPath(notePath);
+          noteService.updateNote(note);
+        });
+    sendListNotesInfo(conn);
   }
 
-  public void removeFolder(final SockMessage fromMessage) throws IOException {
-    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
+  public void removeFolder(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
+    final String folderPath = normalizePath(fromMessage.getNotNull("id")) + "/";
 
-    final String folderPath = "/" + fromMessage.getNotNull("id") + "/";
-
-    final Set<String> admin = new HashSet<>();
-    admin.addAll(Configuration.getAdminUsers());
-    admin.addAll(Configuration.getAdminGroups());
-
-    final Set<String> userRoles = new HashSet<>();
-    userRoles.addAll(authenticationInfo.getRoles());
-    userRoles.add(authenticationInfo.getUser());
-
-    final List<Note> notes = noteService.getAllNotes()
-            .stream()
-            .filter(note -> !note.getPath().startsWith(folderPath))
-            .filter(note -> new HashSet<>(userRoles).removeAll(note.getReaders()) || new HashSet<>(admin).removeAll(userRoles))
-            .collect(Collectors.toList());
-
-    for (final Note note : notes) {
-      noteService.deleteNote(note);
-    }
+    noteService.getAllNotes().stream()
+        .filter(this::userHasOwnerPermission)
+        .filter(note -> note.getPath().startsWith(folderPath))
+        .forEach(noteService::deleteNote);
+    sendListNotesInfo(conn);
   }
 
-  // TODO: ??????
-  public void emptyTrash(final SockMessage fromMessage) throws IOException {
-    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
-
-    final Set<String> admin = new HashSet<>();
-    admin.addAll(Configuration.getAdminUsers());
-    admin.addAll(Configuration.getAdminGroups());
-
-    final Set<String> userRoles = new HashSet<>();
-    userRoles.addAll(authenticationInfo.getRoles());
-    userRoles.add(authenticationInfo.getUser());
-
-    final List<Note> notes = noteService.getAllNotes()
-            .stream()
-            .filter(note -> !note.getPath().startsWith("/" + TRASH_FOLDER + "/"))
-            .filter(note -> new HashSet<>(userRoles).removeAll(note.getReaders()) || new HashSet<>(admin).removeAll(userRoles))
-            .collect(Collectors.toList());
-
-    for (final Note note : notes) {
-      noteService.deleteNote(note);
-    }
+  public void emptyTrash(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
+    noteService.getAllNotes().stream()
+        .filter(note -> note.getPath().startsWith("/" + TRASH_FOLDER + "/"))
+        .filter(this::userHasOwnerPermission)
+        .forEach(noteService::deleteNote);
+    sendListNotesInfo(conn);
   }
 
-  public void restoreAll(final SockMessage fromMessage) throws IOException {
-    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
-
-    final Set<String> admin = new HashSet<>();
-    admin.addAll(Configuration.getAdminUsers());
-    admin.addAll(Configuration.getAdminGroups());
-
-    final Set<String> userRoles = new HashSet<>();
-    userRoles.addAll(authenticationInfo.getRoles());
-    userRoles.add(authenticationInfo.getUser());
-
-    final List<Note> notes = noteService.getAllNotes()
-            .stream()
-            .filter(note -> !note.getPath().startsWith("/" + TRASH_FOLDER + "/"))
-            .filter(note -> new HashSet<>(userRoles).removeAll(note.getReaders()) || new HashSet<>(admin).removeAll(userRoles))
-            .collect(Collectors.toList());
-
-    for (final Note note : notes) {
-      final String notePath = note.getPath().substring(TRASH_FOLDER.length() + 1);
-      note.setPath(notePath);
-      noteService.updateNote(note);
-    }
+  public void restoreAll(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
+    noteService.getAllNotes().stream()
+        .filter(this::userHasOwnerPermission)
+        .filter(note -> note.getPath().startsWith("/" + TRASH_FOLDER + "/"))
+        .forEach(note -> {
+          String notePath = normalizePath(note.getPath().substring(TRASH_FOLDER.length() + 1));
+          note.setPath(notePath);
+          noteService.updateNote(note);
+        });
+    sendListNotesInfo(conn);
   }
 
   public void clearAllParagraphOutput(final WebSocketSession conn, final SockMessage fromMessage) throws IOException {
@@ -397,5 +304,38 @@ public class NoteHandler extends AbstractHandler {
       paragraph.setJobId(null);
       noteService.updateParagraph(note, paragraph);
     }
+  }
+
+  private boolean userHasOwnerPermission(final Note note) {
+    final AuthenticationInfo authenticationInfo = AuthorizationService.getAuthenticationInfo();
+
+    final Set<String> admin = new HashSet<>();
+    admin.addAll(Configuration.getAdminUsers());
+    admin.addAll(Configuration.getAdminGroups());
+
+    final Set<String> userRoles = new HashSet<>();
+    userRoles.addAll(authenticationInfo.getRoles());
+    userRoles.add(authenticationInfo.getUser());
+
+    return userRoles.removeAll(admin) || userRoles.removeAll(note.getOwners());
+  }
+
+  private static String normalizePath(String path) {
+    // fix 'folder/noteName' --> '/folder/noteName'
+    if (!path.startsWith("/")) {
+      path = "/" + path;
+    }
+
+    // fix '///folder//noteName' --> '/folder/noteName'
+    while (path.contains("//")) {
+      path = path.replaceAll("//", "/");
+    }
+
+    //fix '/folder/noteName/' --> '/folder/noteName'
+    if (path.endsWith("/")) {
+      path = path.substring(0, path.length() - 1);
+    }
+
+    return path;
   }
 }
